+++ trunk/engine/src/area.h
@@ -11,20 +11,20 @@ class Area
     public:
         const std::string id;
     protected:
-        int _x;
-        int _y;
-        int _width;
-        int _height;
+        float _x;
+        float _y;
+        float _width;
+        float _height;
         std::string _event;
     public:
         Area(const std::string name);
         virtual ~Area();
     public:
-        void size(const int x, const int y, const int width, const int height);
-        int x() const;
-        int y() const;
-        int w() const;
-        int h() const;
+        void size(const float x, const float y, const float width, const float height);
+        float x() const;
+        float y() const;
+        float w() const;
+        float h() const;
         std::string event() const;
         void setEvent(const std::string event);
 };
